Squashed commit of the getSize-semantics branch
commit 17f95490e12d2c7b376716d4b4287ba4f99fdbc9
Author: Deividas <deividas@deividas-HP-ProBook-450-G1.(none)>
Date:   Mon Dec 8 17:44:54 2014 +0200

    filter size

commit f7834cdc4c68fe9bf39e980aeb8e9a496649e51f
Author: Deividas <deividas@deividas-HP-ProBook-450-G1.(none)>
Date:   Mon Dec 8 17:34:06 2014 +0200

    int instead of size_t to be able to return -1

+++ CollectionAdapter.h
@@ -41,7 +41,7 @@ struct CollectionAdapter {
     static ValueType& getByIndex(ThisCol& c, size_t i);
     static ConstValueType& getByIndex(ConstCol& c, size_t i);
 
-    static size_t size(ConstCol& c);
+    static int size(ConstCol& c);
 
     static void erase(ThisCol& c, Iterator i);
     static void erase(ThisCol& c, Iterator beg, Iterator end);
