+++ HybridVector.h
@@ -308,7 +308,7 @@ struct CollectionAdapter< HybridVector<T,sz,Additional,Alloc> > {
         return c.getByIndex(i);
     }
 
-    static size_t size(ConstCol& c) {
+    static int size(ConstCol& c) {
         return c.size();
     }
 
@@ -450,7 +450,7 @@ struct CollectionAdapter< const HybridVector<T,sz,Additional,Alloc> > {
         return c.getByIndex(i);
     }
 
-    static size_t size(ConstCol& c) {
+    static int size(ConstCol& c) {
         return c.size();
     }
