+++ adapters/StdVector.h
@@ -127,7 +127,7 @@ struct CollectionAdapter< std::vector<T,Alloc<T> > > {
 		return c.cend();
 	}
 
-	static size_t size(ConstCol& c) {
+	static int size(ConstCol& c) {
 		return c.size();
 	}
 
@@ -235,7 +235,7 @@ struct CollectionAdapter< const std::vector<T,Alloc<T> > > {
 		return c.cend();
 	}
 
-	static size_t size(ConstCol& c) {
+	static int size(ConstCol& c) {
 		return c.size();
 	}
