+++ detail/OnceTraversable.h
@@ -237,7 +237,7 @@ struct CollectionAdapter<
     }
 
     template <class U = int>
-    static size_t size(ConstCol& c) {
+    static int size(ConstCol& c) {
         static_assert(templatious::util::DummyResolver<U,false>::val,
                 "OnceTraversable is not full fledged collection"
                 " and can be traversed with iterators only.");
@@ -417,7 +417,7 @@ struct CollectionAdapter<
     }
 
     template <class U = int>
-    static size_t size(ConstCol& c) {
+    static int size(ConstCol& c) {
         static_assert(templatious::util::DummyResolver<U,false>::val,
                 "OnceTraversable is not full fledged collection"
                 " and can be traversed with iterators only.");
@@ -572,7 +572,7 @@ struct CollectionAdapter<
     }
 
     template <class U = int>
-    static size_t size(ConstCol& c) {
+    static int size(ConstCol& c) {
         static_assert(templatious::util::DummyResolver<U,false>::val,
                 "OnceTraversable is not full fledged collection"
                 " and can be traversed with iterators only.");
